+++ benchmarks/benchmark_indicators.py
@@ -138,6 +138,7 @@ INDICATORS: tuple[IndicatorSpec, ...] = (
     IndicatorSpec("KalmanLocalLinearTrend", ("close",), batch_inputs=("close",)),
     IndicatorSpec("KalmanMovingAverage", ("close",), batch_inputs=("close",)),
     IndicatorSpec("KalmanPredictionBands", ("close",), batch_inputs=("close",)),
+    IndicatorSpec("KalmanTrendSignal", ("close",), batch_inputs=("close",)),
     IndicatorSpec("KalmanVelocityOscillator", ("close",), batch_inputs=("close",)),
     IndicatorSpec("Kama", ("close",), batch_inputs=("input",)),
     IndicatorSpec("KeltnerChannel", ("close", "high", "low")),